+++ packages/go-gen/tests/cosmwasm_std__DelegatorWithdrawAddressResponse.go
@@ -1,3 +1,3 @@
 type DelegatorWithdrawAddressResponse struct {
-    WithdrawAddress string `json:"withdraw_address"`
+	WithdrawAddress string `json:"withdraw_address"`
 }