+++ packages/std/src/ibc/callbacks.rs
@@ -4,7 +4,7 @@
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
-use crate::{Addr, IbcAcknowledgement, IbcPacket, Uint64};
+use crate::{Addr, Coin, IbcAcknowledgement, IbcPacket, Uint64};
 
 /// This is just a type representing the data that has to be sent with the IBC message to receive
 /// callbacks. It should be serialized and sent with the IBC message.
@@ -205,6 +205,29 @@ impl IbcTimeoutCallbackMsg {
 pub struct IbcDestinationCallbackMsg {
     pub packet: IbcPacket,
     pub ack: IbcAcknowledgement,
+    /// When the underlying packet is a successful transfer message,
+    /// this field contains information about the transfer. Otherwise it is empty.
+    ///
+    /// This is always empty on chains using CosmWasm < 3.0
+    #[serde(default)]
+    pub transfer: Option<IbcTransferCallback>,
+}
+
+#[derive(
+    Serialize, Deserialize, Clone, Debug, PartialEq, Eq, cw_schema::Schemaifier, JsonSchema,
+)]
+pub struct IbcTransferCallback {
+    /// Address of the sender of the transfer.
+    /// Note that this is *not* a valid address on the destination chain.
+    pub sender: String,
+    /// Address of the receiver of the transfer.
+    /// Since this is on the destination chain, this is a valid address.
+    pub receiver: Addr,
+    /// The funds that were transferred.
+    ///
+    /// When the callback is executed, the transfer is completed already and the coins are now owned
+    /// by the receiver.
+    pub funds: Vec<Coin>,
 }
 
 #[cfg(test)]