+++ packages/std/src/timestamp.rs
@@ -1,5 +1,4 @@
 use core::fmt;
-use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
 use crate::Uint64;