+++ CHANGELOG.md
@@ -33,6 +33,10 @@ and this project adheres to
 - cosmwasm-vm: Add `ibc2_packet_send` entrypoint ([#2477])
 - cosmwasm-std: Add Tx hash to TransactionInfo and make it non exhaustive
   ([#2480])
+- cosmwasm-std: Add `WasmQuery::RawRange` query to allow querying raw storage
+  ranges of different contracts. ([#2471])
+- cosmwasm-std: Add `transfer` field to `IbcDestinationCallbackMsg`, providing
+  an easier way to handle an IBC transfer in a destination callback. ([#2484])
 
 ## Changed
 
@@ -130,11 +134,13 @@ and this project adheres to
 [#2454]: https://github.com/CosmWasm/cosmwasm/pull/2454
 [#2458]: https://github.com/CosmWasm/cosmwasm/pull/2458
 [#2467]: https://github.com/CosmWasm/cosmwasm/pull/2467
+[#2471]: https://github.com/CosmWasm/cosmwasm/pull/2471
 [#2472]: https://github.com/CosmWasm/cosmwasm/pull/2472
 [#2473]: https://github.com/CosmWasm/cosmwasm/pull/2473
 [#2477]: https://github.com/CosmWasm/cosmwasm/pull/2477
 [#2479]: https://github.com/CosmWasm/cosmwasm/pull/2479
 [#2480]: https://github.com/CosmWasm/cosmwasm/pull/2480
+[#2484]: https://github.com/CosmWasm/cosmwasm/pull/2484
 
 ## [2.2.0] - 2024-12-17